+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.9.6
 
+* `OmniHashableMixin` will be faster now